chore: Modify excute absolute file path
